# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second Brain API",
    "description": "Knowledge management and learning system API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/practice/session": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Create Session",
        "description": "Create a new practice session.\n\nReturns a balanced session with:\n- 40% due spaced rep cards\n- 30% weak spot exercises\n- 30% new/interleaved content",
        "operationId": "create_session_api_practice_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/session/{session_id}/end": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "End Session",
        "description": "End a practice session and get summary statistics.",
        "operationId": "end_session_api_practice_session__session_id__end_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/exercise/generate": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Generate Exercise",
        "description": "Generate an adaptive exercise for a topic.\n\nExercise type and difficulty are selected based on mastery level:\n- < 0.3: Worked examples, code completions\n- 0.3-0.7: Free recall, implementations\n- > 0.7: Applications, teach-back",
        "operationId": "generate_exercise_api_practice_exercise_generate_post",
        "parameters": [
          {
            "name": "mastery_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "description": "Current mastery level",
              "default": 0.5,
              "title": "Mastery Level"
            },
            "description": "Current mastery level"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/exercises": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "List Exercises",
        "description": "List all exercises with optional filtering.\n\nReturns exercises ordered by creation date (newest first).",
        "operationId": "list_exercises_api_practice_exercises_get",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by topic path",
              "title": "Topic"
            },
            "description": "Filter by topic path"
          },
          {
            "name": "exercise_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exercise type",
              "title": "Exercise Type"
            },
            "description": "Filter by exercise type"
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseDifficulty"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by difficulty",
              "title": "Difficulty"
            },
            "description": "Filter by difficulty"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum exercises to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum exercises to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  },
                  "title": "Response List Exercises Api Practice Exercises Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/exercise/{exercise_id}": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Get Exercise",
        "description": "Get an exercise by ID.",
        "operationId": "get_exercise_api_practice_exercise__exercise_id__get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/submit": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Submit Attempt",
        "description": "Submit an exercise attempt for evaluation.\n\nReturns detailed feedback including:\n- Score (0-1)\n- Covered and missing key points\n- Misconceptions identified\n- LLM-generated feedback\n- Solution reveal",
        "operationId": "submit_attempt_api_practice_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/attempt/{attempt_id}/confidence": {
      "patch": {
        "tags": [
          "practice"
        ],
        "summary": "Update Confidence",
        "description": "Update confidence rating after viewing feedback.",
        "operationId": "update_confidence_api_practice_attempt__attempt_id__confidence_patch",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptConfidenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Confidence Api Practice Attempt  Attempt Id  Confidence Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptSubmitRequest": {
        "properties": {
          "exercise_id": {
            "type": "integer",
            "title": "Exercise Id",
            "description": "Exercise ID"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response",
            "description": "Text response"
          },
          "response_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Code",
            "description": "Code response"
          },
          "confidence_before": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Before",
            "description": "Self-reported confidence before (1-5)"
          },
          "time_spent_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Spent Seconds"
          }
        },
        "type": "object",
        "required": [
          "exercise_id"
        ],
        "title": "AttemptSubmitRequest",
        "description": "Request to submit an exercise attempt.\n\nSupports both text responses (for conceptual exercises) and code responses\n(for programming exercises). The confidence_before field captures the learner's\nself-assessment prior to receiving feedback, enabling calibration tracking."
      },
...
      "ExerciseDifficulty": {
        "type": "string",
        "enum": [
          "foundational",
          "intermediate",
          "advanced"
        ],
        "title": "ExerciseDifficulty",
        "description": "Difficulty levels aligned with mastery progression.\n\nDifficulty selection is adaptive based on learner mastery:\n- mastery < 0.3: FOUNDATIONAL (worked examples, completions)\n- mastery 0.3-0.7: INTERMEDIATE (free recall, implementations)\n- mastery > 0.7: ADVANCED (applications, refactoring)"
      },
      "ExerciseGenerateRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic",
            "description": "Topic to generate exercise for"
          },
          "exercise_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExerciseType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific type, or auto-select based on mastery"
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExerciseDifficulty"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific difficulty, or auto-select based on mastery"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Programming language for code exercises"
          },
          "source_content_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Content Ids",
            "description": "Optional content IDs to base exercise on"
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "ExerciseGenerateRequest",
        "description": "Request to generate an exercise.\n\nThe LLM exercise generator uses the topic and optional source content to create\na contextually relevant exercise. If type/difficulty are not specified, they are\nautomatically selected based on the learner's current mastery level for the topic."
      },
      "ExerciseResponse": {
        "properties": {
          "exercise_type": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "description": "Topic path (e.g., ml/transformers)"
          },
          "difficulty": {
            "$ref": "#/components/schemas/ExerciseDifficulty",
            "default": "intermediate"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "Main exercise prompt"
          },
          "hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hints",
            "description": "Progressive hints"
          },
          "expected_key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Key Points",
            "description": "Key points for a good answer"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exercise_uuid": {
            "type": "string",
            "title": "Exercise Uuid"
          },
          "worked_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worked Example"
          },
          "follow_up_problem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Problem"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "starter_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Code"
          },
          "buggy_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buggy Code"
          },
          "estimated_time_minutes": {
            "type": "integer",
            "title": "Estimated Time Minutes",
            "default": 10
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "exercise_type",
          "topic",
          "prompt",
          "id",
          "exercise_uuid"
        ],
        "title": "ExerciseResponse",
        "description": "Exercise response (hides solution until attempted).\n\nReturns the exercise prompt and any scaffolding (hints, starter code, buggy code)\nbut withholds the solution until after the learner submits an attempt. This\nprevents premature exposure to answers and encourages genuine retrieval practice."
      },
      "ExerciseType": {
        "type": "string",
        "enum": [
          "free_recall",
          "self_explain",
          "worked_example",
          "application",
          "compare_contrast",
          "teach_back",
          "code_implement",
          "code_complete",
          "code_debug",
          "code_refactor",
          "code_explain"
        ],
        "title": "ExerciseType",
        "description": "Types of exercises for active learning.\n\nText-based exercises:\n- FREE_RECALL: Explain concept from memory\n- SELF_EXPLAIN: Explain why/how something works\n- WORKED_EXAMPLE: Step-by-step solution followed by similar problem\n- APPLICATION: Apply concept to novel situation\n- COMPARE_CONTRAST: Compare two concepts or approaches\n- TEACH_BACK: Explain as if teaching someone (Feynman technique)\n\nCode-based exercises:\n- CODE_IMPLEMENT: Write code from scratch\n- CODE_COMPLETE: Fill in blanks in code\n- CODE_DEBUG: Find and fix bugs\n- CODE_REFACTOR: Improve existing code\n- CODE_EXPLAIN: Explain what code does"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionCreateRequest": {
        "properties": {
          "duration_minutes": {
            "type": "integer",
            "maximum": 120,
            "minimum": 5,
            "title": "Duration Minutes",
            "description": "Target duration",
            "default": 15
          },
          "topic_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Filter",
            "description": "Optional topic to focus on"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType",
            "default": "practice"
          },
          "reuse_exercises": {
            "type": "boolean",
            "title": "Reuse Exercises",
            "description": "Reuse existing exercises instead of generating new ones. Set to False to always generate fresh exercises via LLM.",
            "default": true
          }
        },
        "type": "object",
        "title": "SessionCreateRequest",
        "description": "Request to create a practice session.\n\nCreates a balanced session mixing due spaced rep cards (40%), weak spot exercises\n(30%), and new content (30%). The topic_filter focuses the session on a specific\narea, otherwise content is drawn from across all learned topics.\n\nWhen reuse_exercises=True, existing exercises for the topic are used instead of\ngenerating new ones via LLM. This is faster and free (no API calls)."
      },
...
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "estimated_duration_minutes": {
            "type": "number",
            "title": "Estimated Duration Minutes"
          },
          "topics_covered": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics Covered"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "items",
          "estimated_duration_minutes",
          "topics_covered",
          "session_type"
        ],
        "title": "SessionResponse",
        "description": "Practice session with items.\n\nContains the ordered list of items to complete during the session. Items are\ninterleaved for optimal learning with worked examples placed first. The\nestimated duration helps users plan their available study time."
      },
      "SessionSummary": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "duration_minutes": {
            "type": "number",
            "title": "Duration Minutes"
          },
          "cards_reviewed": {
            "type": "integer",
            "title": "Cards Reviewed"
          },
          "exercises_completed": {
            "type": "integer",
            "title": "Exercises Completed"
          },
          "correct_count": {
            "type": "integer",
            "title": "Correct Count"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "average_score": {
            "type": "number",
            "title": "Average Score"
          },
          "mastery_changes": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Mastery Changes",
            "description": "Topic -> mastery delta"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "duration_minutes",
          "cards_reviewed",
          "exercises_completed",
          "correct_count",
          "total_count",
          "average_score"
        ],
        "title": "SessionSummary",
        "description": "Summary after completing a session.\n\nProvides performance metrics and mastery changes from the session. The\nmastery_changes dict shows how each practiced topic's mastery score changed,\nhelping learners see the impact of their study session."
      },
      "SessionType": {
        "type": "string",
        "enum": [
          "review",
          "practice",
          "focused",
          "weak_spots"
        ],
        "title": "SessionType",
        "description": "Types of practice sessions."
      },
...
    }
  }
}